requirements.txt
SQLAlchemy==2.0.32
python-dotenv==1.0.1
PyMySQL==1.1.1
cryptography==43.0.0
boto3==1.35.9